Rework forum/main.js and add new LandingPage component
Hooking into the IndexPage component and app.routes.index turned out to be not-so-good when using the Tags extension as the Home Page, and still allowed logged-out users to view discussion pages.

Extending each of the Page components view decouples the extension from the IndexPage component, which meant the landing page would only work when the sites Home Page was set to 'All Discussions'. Using the old method didnt allow us to make the site 'membership-only' in a sense as the endpoints were still reachable for logged out users.

Resolves: #3